requirements.txt
sentence-transformers==4.1.0
langchain
flask
pypdf
python-dotenv
pinecone-client
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain-google-genai
-e .
